First Initial Project Upload. LgwueJniHtRzhtUQcxfxf9ZHvAH42HsVrA Donate if helpful.
Python-Flask-Crypto Explorer !!!

This is a small weight Python-Flask-Crypto Explorer for such Coins as PRUX-Coin, BTC, LTC, DOGE, BCH, EFL, XVG, PRUX   and more!

Designed for my Coin PRUX, may need some RPC command corrections for using at last for the other similar Coins. For DB we use actually sqlite3,  was easyier for the start of coding,  may the selection of this db is not the best, also need changes may in Future.  

Fell free to use and improve my Project its running but still not finsihed 100%. 

If you find it very helpful, I would be happy for a donation so that I know that I have done something good and will do it again in the future, then please donate to my LTC address,  LTC-->    LgwueJniHtRzhtUQcxfxf9ZHvAH42HsVrA